# Fix various performance issues