I'm investigating some performance problems with a facility location problem: https://github.com/jump-dev/JuMPPaperBenchmarks/blob/main/benchmark/facility.jl

```Julia
function _benchmark(n)
    model = Model(Gurobi.Optimizer)
    set_silent(model)
    solve_facility(model, n, n)
end
```

Before
```julia
julia> @time _benchmark(60);
Academic license - for non-commercial use only - expires 2022-06-19
  9.661598 seconds (57.29 M allocations: 4.051 GiB, 28.08% gc time)

julia> @time _benchmark(60);
Academic license - for non-commercial use only - expires 2022-06-19
  8.932473 seconds (57.29 M allocations: 4.051 GiB, 23.78% gc time)
```
After
```julia
julia> @time _benchmark(60);
Academic license - for non-commercial use only - expires 2022-06-19
  7.956889 seconds (47.67 M allocations: 3.210 GiB, 25.95% gc time)

julia> @time _benchmark(60);
Academic license - for non-commercial use only - expires 2022-06-19
  7.773976 seconds (47.67 M allocations: 3.210 GiB, 22.78% gc time)
```